+++ sync/SyncDataCollection.ts
@@ -1,5 +1,5 @@
 ﻿import Arrays = require("../../ts-mortar/utils/Arrays");
-import Defer = require("../../ts-mortar/promises/Defer");
+import Defer = require("../../ts-promise-tasks/promises/Defer");
 
 /** Combines functionality for two operations in one class:
  *  - Sync a local data collection to a remote data collection (refered to as 'syncing up').
@@ -128,12 +128,12 @@ class SyncDataCollection {
 
             return syncSetting.syncUpFunc(params, data).then(function (res) {
                 return res;
-            }, function (err) {
-                return {
+            }, function (err): SyncError {
+                throw {
                     collectionName: localColl.getName(),
-                    syncingToServer: true,
+                    syncingUp: true,
                     error: err
-                }
+                };
             });
         });
     }